Remove hard coded objects and refactor converters
