+++ src/main/java/org/venice/beachfront/bfapi/model/Scene.java
@@ -9,6 +9,7 @@ import javax.persistence.Transient;
 
 import org.hibernate.annotations.Type;
 import org.joda.time.DateTime;
+import org.venice.beachfront.bfapi.model.converter.TimestampConverter;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.vividsolutions.jts.geom.Geometry;
